fix (search) : refactor search type selector components
