+++ components/search/typeSelector/BoardDialog.vue
